+++ front-page.php
@@ -161,21 +161,21 @@ $category_emploi = get_category_by_slug( 'offres-emploi' );
 									'icon' => 'news',
 									'color' => '#009fb8',
 									'title' => "110 pays",
-									'text' => "Le réseau Tela Botanica rassemble des passionnés de tous pays",
+									'text' => __("Le réseau Tela Botanica rassemble des passionnés de tous pays", 'telabotanica'),
 								],
 								[
 									'href' => '#', // TODO
 									'icon' => 'members',
 									'color' => '#ff5d55',
 									'title' => "28 214 telabotanistes",
-									'text' => "Trouvez les inscrits près de chez vous et prenez contact",
+									'text' => __("Trouvez les inscrits près de chez vous et prenez contact", 'telabotanica'),
 								],
 								[
 									'href' => '#', // TODO
 									'icon' => 'home',
 									'color' => '#918a6f',
 									'title' => "172 structures",
-									'text' => "À travers le monde, les structures locales vous accompagnent dans vos découvertes naturalistes",
+									'text' => __("À travers le monde, les structures locales vous accompagnent dans vos découvertes naturalistes", 'telabotanica'),
 								]
 							],
 							'modifiers' => 'layout-column-item'
@@ -192,17 +192,37 @@ $category_emploi = get_category_by_slug( 'offres-emploi' );
 			</div>
 
 			<?php
+			the_telabotanica_block('contribute', [
+				'background_color' => '#f8f5ef',
+				'title' => __('Contribuez dès maintenant', 'telabotanica'),
+				'query' => new WP_Query([
+					'post_type' => 'tb_participer',
+					'posts_per_page' => 3,
+					'orderby' => 'rand'
+				]),
+				'buttons' => [
+					'display' => 'buttons',
+					'items' => [
+						[
+							'href' => get_permalink( get_page_by_path( 'comment-participer' ) ),
+							'text' => __('Trouver un moyen de participer', 'telabotanica'),
+							'modifiers' => 'orange'
+						]
+					]
+				]
+			]);
+
 			the_telabotanica_block('focus', [
 				'background_color' => '#f8f5ef',
-				'title' => 'Rejoignez un projet citoyen',
-				'intro' => 'Un programme de <strong>science participative</strong> est un programme conduit en partenariat entre des observateurs (citoyens) et un laboratoire ou une structure à vocation scientifique, visant à observer ou étudier un phénomène dans le cadre d’un protocole bien défini.',
+				'title' => __('Rejoignez un projet citoyen', 'telabotanica'),
+				'intro' => __('Un programme de <strong>science participative</strong> est un programme conduit en partenariat entre des observateurs (citoyens) et un laboratoire ou une structure à vocation scientifique, visant à observer ou étudier un phénomène dans le cadre d’un protocole bien défini.', 'telabotanica'),
 				'display_buttons' => ['intro'],
 				'intro_buttons' => [
 					'display' => 'buttons',
 					'items' => [
 						[
 							'href' => get_permalink( get_page_by_path( 'projets' ) ),
-							'text' => 'Voir tous les projets'
+							'text' => __('Voir tous les projets', 'telabotanica')
 						]
 					]
 				]