+++ inc/styleguide.php
@@ -60,6 +60,7 @@ array_walk($telabotanica_modules, function ($module) {
  * Liste de tous les composants accessibles depuis les fonctions du styleguide
  */
 $telabotanica_blocks = [
+	'contribute',
 	'focus',
 	'list-features',
 	'list-projects',