+++ CHANGELOG.md
@@ -11,6 +11,7 @@ plus de détails sur http://semver.org/lang/fr/
 
 ## x.β.α (bientôt)
 
+* bloc `contribute` ajouté
 * bloc `list-projects` ajouté
 * ajout d'une dépendance (`pug-php`) avec Composer
 * mise à jour du `README`